requirements.txt
maxapi>=0.9.13
pandas>=2.0.0
openpyxl>=3.1.0
python-Levenshtein>=0.23.0
flask>=2.3.0
gunicorn>=20.1.0
python-dotenv>=1.0.0
aiohttp>=3.8.0